pyproject.toml
[project]
name = "mlops-course"
description = "MLOps with Databricks course"
requires-python = ">=3.12, <3.13"
dependencies = [
    "mlflow==3.1.1",
    "cffi==1.17.1",
    "cloudpickle==3.1.1",
    "numpy==1.26.4",
    "pandas==2.3.0",
    "pyarrow==17.0.0",
    "scikit-learn==1.7.0",
    "lightgbm==4.6.0",
    "xgboost==2.0.3",
    "scipy==1.16.0",
    "databricks-sdk==0.62.0",
    "databricks-feature-engineering==0.13.0",
    "databricks-feature-lookup==1.8.0",
    "pydantic==2.11.7",
    "loguru==0.7.3",
    "python-dotenv==1.1.1",
    "pyyaml==6.0.1",
    "category-encoders==2.6.3",
    "matplotlib==3.8.2",
    "seaborn==0.13.2",
    "pyspark>=3.4.0",
    "ray[tune]==2.47.1",
    "optuna==4.4.0",
    "sdv==1.24.0",
    "boto3>=1.40.27",
]

dynamic = ['version']

[project.optional-dependencies]
dev = ["databricks-connect>=16.0, <17",
       "ipykernel>=6.29.5, <7",
       "pip>=25.0.1, <26",
       "pre-commit>=4.1.0, <5",
       "black>=24.0.0",
       "isort>=5.13.0"]

test = ["pre-commit>=4.1.0, <5",
       "pytest>=8.0.0, <9",
       "pytest-cov>=5.0.0, <6",
       "pyspark>=3.4.0, <4",
       "pydantic-settings>=2.0.0"]

ci = ["pre-commit>=4.1.0, <5",
      "pytest>=8.0.0, <9",
      "pytest-cov>=5.0.0, <6",
      "pyspark>=3.4.0, <4"]


ml_advanced = [
    "ray[tune]>=2.10.0",
    "optuna>=3.5.0",
    "hyperopt>=0.2.7",
    "boto3>=1.34.0",
]      

[build-system]
requires = [
  "setuptools>=72.0"]
build-backend = "setuptools.build_meta"


[tool.setuptools.packages.find]
where = ["src"]
include = ["financial_complaints*"]


[tool.setuptools.dynamic]
version = { file = "version.txt" }

[tool.pytest.ini_options]
testpaths = [ "tests" ]
pythonpath = [ "." ]
markers = [
    "ci_exclude: Tests excluded from CI/CD pipeline"
]
addopts = [
    "--cov=src",
    "--cov-report=term",
    "--cov-report=xml:coverage.xml"
]


[tool.ruff]
line-length = 120
lint.select = [
    "F",    # pyflakes rules
    "E",    # pycodestyle error rules
    "W",    # pycodestyle warning rules
    "B",    # flake8-bugbear rules
    "I",    # isort rules
    "UP",   # pyupgrade rules
    "SIM",  # flake8-simplify rules
    "ERA",  # eradicate rules
    "C",    # pylint convention rules
    "D",    # pydocstyle rules
    "ANN"   # flake8-annotations rules
]
lint.ignore = [
    "D107", # Missing docstring in `__init__`
    "D203", # incorrect-blank-line-before-class
    "D213", # multi-line-summary-second-line
    "E501", # line-too-long
    "C901", # too complex
    "ERA001", # commented out code
    "D100", # missing docstring
    "ANN101", # missing type annotation for self
    "ANN102", # missing type annotation for cls
    "ANN401", # dynamically typed expressions
]

[tool.ruff.format]
indent-style = "space"
docstring-code-format = true

[tool.black]
line-length = 120
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 120